--- Data/TestModule.py
@@ -1,5 +1,0 @@
-def __init__(self):
-    print("Initalized")
-
-def TMAltair():
-    print("Test passed!")
